moved polkadot as its own module inside network
