+++ src/network/network.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
-import { PolkadotService } from './polkadot/polkadot.service';
 import { NetworkController } from './controller/network.controller';
 import { NetworkService } from './network.service';
+import { PolkadotModule } from './polkadot/polkadot.module';
 
 @Module({
-  providers: [PolkadotService, NetworkService],
+  imports: [PolkadotModule],
+  providers: [NetworkService],
   controllers: [NetworkController],
 })
 export class NetworkModule {}
