+++ src/network/network.service.ts
@@ -1,19 +1,21 @@
-import { Inject, Injectable } from "@nestjs/common";
-import { WalletBalance } from "./polkadot/dto/WalletBalance.dto";
-import { PolkadotService } from "./polkadot/polkadot.service";
-import { SupportedNetworkNames } from "./supported_network";
+import { Injectable } from '@nestjs/common';
+import { WalletBalance } from './polkadot/dto/WalletBalance.dto';
+import { PolkadotService } from './polkadot/services/polkadot.service';
+import { SupportedNetworkNames } from './supported_network';
 
 @Injectable()
-export class NetworkService{
+export class NetworkService {
+  constructor(private polkadotService: PolkadotService) {}
 
-    constructor(private polkadotService:PolkadotService) {}
-
-    async getBalance(address:string, network:SupportedNetworkNames): Promise<WalletBalance> {
-        switch(network){
-            case SupportedNetworkNames.POLKADOT:
-                return await this.polkadotService.getBalance(address);
-            default:
-                throw new Error(`Network ${network} is not supported`);
-        }
+  async getBalance(
+    address: string,
+    network: SupportedNetworkNames,
+  ): Promise<WalletBalance> {
+    switch (network) {
+      case SupportedNetworkNames.POLKADOT:
+        return await this.polkadotService.getBalance(address);
+      default:
+        throw new Error(`Network ${network} is not supported`);
     }
+  }
 }
