+++ src/network/polkadot/polkadot.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { PolkadotService } from './services/polkadot.service';
+
+@Module({
+  exports: [PolkadotService],
+  providers: [PolkadotService],
+})
+export class PolkadotModule {}
